refactor(test): split host.spec.ts (1190 lines) by concern
共有モックビルダー / FakeWindow を host.test-helpers.ts へ抽出し、
機能ごとに 4 ファイルへ分割:
- host.routing.spec.ts (getPublicKey / getRelays / signEvent ルーティング)
- host.errors.spec.ts (エラー処理・可視性・ライフサイクル)
- host.crypto.spec.ts (nip44 / nip04 encrypt / decrypt)
- host.rate-limiting.spec.ts (per-origin consent rate limiting)

テスト件数は不変。
